+++ lambdas/redis_sync/src/transform_map.py
@@ -1,5 +1,5 @@
 from constants import RedisCacheKey
-from clients import logger
+from common.clients import logger
 from transform_configs import transform_vaccine_map, transform_supplier_permissions
 '''
 Transform config file to format required in REDIS cache.
