+++ lambdas/shared/src/common/clients.py
@@ -1,7 +1,7 @@
 import os
 import logging
-from boto3 import client as boto3_client
-import boto3
+import redis
+from boto3 import client as boto3_client, resource as boto3_resource
 
 logging.basicConfig(level="INFO")
 logger = logging.getLogger()
@@ -12,9 +12,15 @@ CONFIG_BUCKET_NAME = os.getenv("CONFIG_BUCKET_NAME", "variconfig-bucketable-not-
 
 REGION_NAME = os.getenv("AWS_REGION", "eu-west-2")
 
+REDIS_HOST = os.getenv("REDIS_HOST", "")
+REDIS_PORT = os.getenv("REDIS_PORT", 6379)
+
 s3_client = boto3_client("s3", region_name=REGION_NAME)
 firehose_client = boto3_client("firehose", region_name=REGION_NAME)
 
 secrets_manager_client = boto3_client("secretsmanager", region_name=REGION_NAME)
-dynamodb_resource = boto3.resource("dynamodb", region_name=REGION_NAME)
+dynamodb_resource = boto3_resource("dynamodb", region_name=REGION_NAME)
 dynamodb_client = boto3_client("dynamodb", region_name=REGION_NAME)
+
+logger.info(f"Connecting to Redis at {REDIS_HOST}:{REDIS_PORT}")
+redis_client = redis.StrictRedis(host=REDIS_HOST, port=REDIS_PORT, decode_responses=True)
